requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
buildozer==0.39
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Django==2.2.5
Flask==1.1.1
future==0.17.1
google-images-download==2.8.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pdftables-api==1.1.0
pefile==2019.4.18
pexpect==4.7.0
Pillow==6.1.0
ptyprocess==0.6.0
py2exe==0.9.2.2
pycparser==2.19
pygame==1.9.6
PyInstaller==3.5
pyTelegramBotAPI==3.6.6
python-telegram-bot==12.0.0
pytube==9.5.2
pytz==2019.2
pywin32-ctypes==0.2.0
requests==2.22.0
selenium==3.141.0
sh==1.12.14
six==1.12.0
sqlparse==0.3.0
tornado==6.0.3
UNKNOWN==0.0.0
urllib3==1.25.6
virtualenv==16.7.5
Werkzeug==0.16.0